app/src/test/java/com/anggiat/jetpackmoviecatalogue/ui/favorite/movie/FavoriteMovieViewModelTest.java
package com.anggiat.jetpackmoviecatalogue.ui.favorite.movie;

import androidx.arch.core.executor.testing.InstantTaskExecutorRule;
import androidx.lifecycle.MutableLiveData;
import androidx.lifecycle.Observer;
import androidx.paging.PagedList;

import com.anggiat.jetpackmoviecatalogue.data.source.MovieRepository;
import com.anggiat.jetpackmoviecatalogue.data.source.local.entity.MovieEntity;

import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

import java.util.List;

import static org.junit.Assert.*;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

@RunWith(MockitoJUnitRunner.class)
public class FavoriteMovieViewModelTest {
    private FavoriteMovieViewModel viewModel;

    @Rule
    public InstantTaskExecutorRule instantTaskExecutorRule = new InstantTaskExecutorRule();

    @Mock
    private MovieRepository movieRepository;

    @Mock
    private Observer<PagedList<MovieEntity>> observer;

    @Mock
    private PagedList<MovieEntity> pagedList;

    @Before
    public void setUp() {
        viewModel = new FavoriteMovieViewModel(movieRepository);
    }

    @Test
    public void getFavoriteMovie() {
        PagedList<MovieEntity> dummyMovies = pagedList;
        when(dummyMovies.size()).thenReturn(10);
        MutableLiveData<PagedList<MovieEntity>> movies = new MutableLiveData<>();
        movies.setValue(dummyMovies);

        when(movieRepository.getFavoriteMovie()).thenReturn(movies);
        List<MovieEntity> movieEntities = viewModel.getFavoriteMovies().getValue();
        verify(movieRepository).getFavoriteMovie();
        assertNotNull(movieEntities);
        assertEquals(10, movieEntities.size());

        viewModel.getFavoriteMovies().observeForever(observer);
        verify(observer).onChanged(dummyMovies);
    }

}